+++ src/api.rs
@@ -13,7 +13,6 @@ use crate::config::ApiConfig;
 
 pub(crate) struct ApiClient {
     pub base_url: String,
-    pub api_token: String,
     pub client_token: String,
     agent: Agent,
 }
@@ -42,7 +41,6 @@ impl ApiClient {
     pub(crate) fn new(config: &ApiConfig) -> Self {
         Self {
             base_url: config.base_url.to_owned(),
-            api_token: config.api_token.to_owned(),
             client_token: config.client_token.to_owned(),
             agent: Agent::config_builder()
                 .timeout_global(Some(Duration::from_secs(config.timeout_in_seconds)))