+++ CHANGELOG.md
@@ -19,7 +19,7 @@
 - Remove file extension from both hard-coded and configured sound names.
   It will be added by the audio player.
 
-- Fixed outdated property name in examplary configuration file.
+- Fixed outdated property name in exemplary configuration file.
 
 
 ## 0.6.1 (2025-05-02)